move staticmethods to helper module
